channel should be closed if the result of fn-call is null.
core.async channel don't accept nil because nil means the channel
already closed.
In other word, we should close the channel for returning nil.
the channel we are handling in this code is a channel accepting a
result of fn-call. so the channel always must be closed after the call.