resolve all constructor symbols in config when loading time.
All keys which might contains a symbol or string of fn name should
be resolved (by executing the fn) before the use of config object.
Because it is verbose if users need to check if the value of config
could be a symbol, resolve it and execute it everytime when using
a config object.