new configuration option: request-wrapper
Users can assign a function to :request-wrapper key of a configuration
map. the function receive a configuration map and must return a function
which handle a message a service received and return a new (or same)
message.
the :request-wrapper function is called before a received message is
handle by a service.